[C#] Fix bool passing
It turns out that C# always expects bools to be passed as 4-byte
ints, which matches the classic C definition (which was usually
typedef'd to int) but is platform-dependent. On some platforms we
were building with 1-byte ints which led to bools getting mangled
when passing them to C#.

Luckily we already have bools passed to C# separated from bools
passed to Rust, so we can simply redefine `jboolean` to `int32_t`.