Be more conservative about holding a ref to a cloned object
If we clone an object before passing to/from native code, we
generally don't need to hold a reference to that object in the
struct that owns the method.

The one exception we include here is if the object being passed is
either a trait or holds a reference to a trait. Its not clear if
this is required but it seems like a potential sharp edge.