Avoid allocating for byte vecs
Rust's memory allocator doesn't allocate for 0-byte allocations.
However, standard `libc` does. As a result, Rust also doesn't free
0-byte allocations whereas `libc`-based C code will rely on such
allocation being free'd.

This breaks our bindings somewhat as we may allocate for 0-byte
strings or arrays, but once we pass them to Rust they will not be
freed.

Here we skip such allocations when we pass byte arrays (in the form
of u8-vecs, as well as a few other types), avoiding leaks when
empty arrays are passed.